pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "predict_llm"
version = "0.1.0"
authors = [
    {name = "Zilinghan Li"},
    {email = "zilinghan.li@anl.gov"},
]
description = "predict_llm example package."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
    "requests",
    "torch",
    "magic-pdf[full] @ https://wheels.myhloli.com/simple",
    "huggingface_hub",
    "pyserini==0.22.1",
    "tiktoken==0.6.0",
    "faiss-cpu==1.7.4",
    "python-liquid==1.10.2",
    "chonkie[all]",
    "tqdm==4.66.1",
    "transformers==4.44.2",
    "sentence_transformers==2.2.2",
    "pyserini==0.22.1",
    "huggingface_hub"
]

[project.urls]
homepage = "https://github.com/Zilinghan/predict-llm"
documentation = "https://predict_llm.readthedocs.io"
repository = "https://github.com/Zilinghan/predict-llm"

[project.optional-dependencies]
dev = [
    "covdefaults>=2.2",
    "coverage",
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "ruff>=0.2.0",
    "tox",
    "virtualenv",
]
docs = [
    "black==24.4.2",
    "mkdocs-click==0.8.1",
    "mkdocs-gen-files==0.5.0",
    "mkdocs-literate-nav==0.6.1",
    "mkdocs-material==9.5.32",
    "mkdocs-section-index==0.3.9",
    "mkdocstrings==0.25.2",
    "mkdocstrings-python==1.10.8",
    "mike==2.1.3",
]

[tool.codespell]
skip = """
.git,
.github,
__pycache__,
build,
dist,
.*egg-info
"""

[tool.coverage.run]
plugins = ["covdefaults"]
omit = ["examples"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "testing.*"
allow_incomplete_defs = true
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
allow_incomplete_defs = true
allow_untyped_defs = true

[tool.ruff]
line-length = 79
target-version = "py39"

[tool.ruff.format]
indent-style = "space"
quote-style = "single"

[tool.ruff.lint]
# See all rules here: https://beta.ruff.rs/docs/rules
select = [
    # pyflakes
    "F",
    # pycodestyle
    "E",
    # mccabe
    "C90",
    # isort
    "I",
    # pep8-naming
    "N",
    # pydocstyle
    "D",
    # pyupgrade
    "UP",
    # flake8-2020
    "YTT",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-commas
    "COM",
    # flake8-comprehensions
    "C4",
    # flake8-implicit-str-concat
    "ISC",
    # flake8-pytest-style
    "PT",
    # flake8-quotes
    "Q",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # PyLint
    "PL",
    # ruff-specific
    "RUF",
]
extend-ignore = [
    "COM812",  # Trailing comma missing
    "Q001",    # Multiline quotes conflict
]

[tool.ruff.lint.flake8-pytest-style]
parametrize-values-type = "tuple"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "single"

[tool.ruff.lint.isort]
force-single-line = true
known-first-party = ["predict_llm", "test", "testing"]
order-by-type = false
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["F401"]
"*/*_test.py" = ["D10"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.setuptools.packages.find]
exclude = ["tests*", "testing*"]
namespaces = false
